unit testing, echo update to v4

+++ server/router/router.go
@@ -3,18 +3,19 @@ package router
 import (
 	"os"
 	"strings"
+	"flag"
 
 	"github.com/go-playground/validator/v10"
-	"github.com/labstack/echo"
-	"github.com/labstack/echo/middleware"
+	"github.com/labstack/echo/v4"
+	"github.com/labstack/echo/v4/middleware"
 
 	"spiros/db"
 	"spiros/models"
 
-	echoSwagger "github.com/swaggo/echo-swagger"
-
 	// import swagger api documentation drive
 	_ "spiros/docs"
+
+	echoSwagger "github.com/swaggo/echo-swagger"
 )
 
 // CustomValidator type for json body validation
@@ -65,6 +66,10 @@ func NewRouter() *echo.Echo {
 }
 
 func generateRoutePermissions(e *echo.Echo) {
+	if flag.Lookup("test.v") != nil {
+		return
+	}
+
 	routes := e.Routes()
 	err := db.DB.Exec("TRUNCATE TABLE permissions RESTART IDENTITY CASCADE").Error
 	if err != nil {
